[#506] New Event notifications are not sent to Group members  (#511)
* New Event notifications are not sent to Group members [#506]

* Add missed notification
