+++ core/application.go
@@ -125,195 +125,6 @@ func (app *Application) startScheduledPostTask() {
 	log.Printf("successful running of post scheduling task")
 }
 
-func (app *Application) createCalendarEventForGroups(clientID string, adminIdentifiers []model.AccountIdentifiers, current *model.User, event map[string]interface{}, groupIDs []string) (map[string]interface{}, []string, error) {
-	memberships, err := app.findGroupMemberships(clientID, model.MembershipFilter{
-		GroupIDs: groupIDs,
-		UserID:   &current.ID,
-		Statuses: []string{"admin"},
-	})
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if memberships.GetMembershipByAccountID(current.ID) != nil {
-		currentAccount := model.AccountIdentifiers{AccountID: &current.ID, ExternalID: &current.NetID}
-		createdEvent, err := app.calendar.CreateCalendarEvent(adminIdentifiers, currentAccount, event, current.OrgID, current.AppID, groupIDs)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		if createdEvent != nil {
-			var mappedGroupIDs []string
-			eventID := createdEvent["id"].(string)
-
-			for _, membership := range memberships.Items {
-				mapping, err := app.storage.CreateEvent(clientID, eventID, membership.GroupID, nil, &model.Creator{
-					UserID: current.ID,
-					Name:   current.Name,
-					Email:  current.Email,
-				})
-				if err != nil {
-					log.Printf("Error create goup mapping: %s", err)
-				}
-				if mapping != nil {
-					mappedGroupIDs = append(mappedGroupIDs, mapping.GroupID)
-				}
-			}
-			return createdEvent, mappedGroupIDs, nil
-		}
-	}
-
-	return nil, nil, nil
-}
-
-func (app *Application) createCalendarEventForGroupsMembers(clientID string, orgID string, appID string, eventID string, groupIDs []string, members []model.ToMember) error {
-	for _, groupID := range groupIDs {
-
-		var userIDs []string
-		memberships, err := app.findGroupMemberships(clientID, model.MembershipFilter{
-			GroupIDs: []string{groupID},
-			Statuses: []string{"admin", "member"},
-		})
-		if err != nil {
-			return err
-		}
-
-		for _, membership := range memberships.Items {
-			if len(members) > 0 {
-				for _, toMember := range members {
-					if toMember.UserID == membership.UserID {
-						userIDs = append(userIDs, membership.UserID)
-						break
-					}
-				}
-			} else {
-				userIDs = append(userIDs, membership.UserID)
-			}
-		}
-		err = app.calendar.AddPeopleToCalendarEvent(userIDs, eventID, orgID, appID)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (app *Application) createCalendarEventSingleGroup(clientID string, current *model.User, event map[string]interface{}, groupID string, members []model.ToMember) (map[string]interface{}, []model.ToMember, error) {
-	memberships, err := app.findGroupMemberships(clientID, model.MembershipFilter{
-		GroupIDs: []string{groupID},
-		UserID:   &current.ID,
-		Statuses: []string{"admin"},
-	})
-	if err != nil {
-		return nil, nil, err
-	}
-
-	var groupIDs []string
-	if groupID != "" {
-		groupIDs = append(groupIDs, groupID)
-	}
-
-	if memberships.GetMembershipByAccountID(current.ID) != nil {
-		currentAccount := model.AccountIdentifiers{AccountID: &current.ID, ExternalID: &current.ExternalID}
-		createdEvent, err := app.calendar.CreateCalendarEvent([]model.AccountIdentifiers{}, currentAccount, event, current.OrgID, current.AppID, groupIDs)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		if createdEvent != nil {
-			var mappedGroupIDs []string
-			eventID := createdEvent["id"].(string)
-
-			mapping, err := app.storage.CreateEvent(clientID, eventID, groupID, members, &model.Creator{
-				UserID: current.ID,
-				Name:   current.Name,
-				Email:  current.Email,
-			})
-			if err != nil {
-				log.Printf("Error create goup mapping: %s", err)
-			}
-			if mapping != nil {
-				mappedGroupIDs = append(mappedGroupIDs, mapping.GroupID)
-			}
-
-			return createdEvent, members, nil
-		}
-	}
-
-	return nil, nil, nil
-}
-
-func (app *Application) updateCalendarEventSingleGroup(clientID string, current *model.User, event map[string]interface{}, groupID string, members []model.ToMember) (map[string]interface{}, []model.ToMember, error) {
-	memberships, err := app.findGroupMemberships(clientID, model.MembershipFilter{
-		GroupIDs: []string{groupID},
-		UserID:   &current.ID,
-		Statuses: []string{"admin"},
-	})
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if len(memberships.Items) > 0 {
-		var groupIDs []string
-		for _, membership := range memberships.Items {
-			groupIDs = append(groupIDs, membership.GroupID)
-		}
-
-		eventID := event["id"].(string)
-		currentAccount := model.AccountIdentifiers{AccountID: &current.ID, ExternalID: &current.NetID}
-		createdEvent, err := app.calendar.UpdateCalendarEvent(currentAccount, eventID, event, current.OrgID, current.AppID)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		if createdEvent != nil {
-			var mappedGroupIDs []string
-			eventID := createdEvent["id"].(string)
-
-			err := app.storage.UpdateEvent(clientID, eventID, groupID, members)
-			if err != nil {
-				return nil, nil, err
-			}
-
-			for _, groupID := range groupIDs {
-				mapping, err := app.storage.CreateEvent(clientID, eventID, groupID, members, &model.Creator{
-					UserID: current.ID,
-					Name:   current.Name,
-					Email:  current.Email,
-				})
-				if err != nil {
-					log.Printf("app.updateCalendarEventSingleGroup() Error create goup mapping: %s", err)
-				}
-				if mapping != nil {
-					mappedGroupIDs = append(mappedGroupIDs, mapping.GroupID)
-				}
-			}
-			return createdEvent, members, nil
-		}
-	}
-
-	return nil, nil, nil
-}
-
-func (app *Application) getGroupCalendarEvents(clientID string, current *model.User, groupID string, published *bool, filter model.GroupEventFilter) (map[string]interface{}, error) {
-	mappings, err := app.storage.FindEvents(clientID, current, groupID, true)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(mappings) > 0 {
-		var eventIDs []string
-		for _, eventMapping := range mappings {
-			eventIDs = append(eventIDs, eventMapping.EventID)
-		}
-		currentAccount := model.AccountIdentifiers{AccountID: &current.ID, ExternalID: &current.NetID}
-		return app.calendar.GetGroupCalendarEvents(currentAccount, eventIDs, current.AppID, current.OrgID, published, filter)
-	}
-
-	return nil, err
-}
-
 // NewApplication creates new Application
 func NewApplication(version string, build string, storage Storage, notifications Notifications, authman Authman, core *corebb.Adapter,
 	rewards *rewards.Adapter, calendar *calendar.Adapter, serviceID string, logger *logs.Logger, config *model.ApplicationConfig) *Application {
